Relocate Images.kt to the component package.
