+++ app/src/main/java/com/OxGames/Pluvia/ui/screen/library/components/LibraryAppItem.kt
@@ -17,9 +17,9 @@ import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import com.OxGames.Pluvia.data.LibraryItem
+import com.OxGames.Pluvia.ui.component.ListItemImage
 import com.OxGames.Pluvia.ui.internal.fakeAppInfo
 import com.OxGames.Pluvia.ui.theme.PluviaTheme
-import com.OxGames.Pluvia.ui.util.ListItemImage
 
 @Composable
 internal fun AppItem(
